fixup: rename file, call the function
